Switch back to unserializing on Ruby side
With the Quassel custom types registered in the C++ extension, The
Ruby code can unserialize their objects. The added qt_debug helper
shows it since QDebug outputs the same for these variants as it did
in the C++ extension. Attempting to convert them with Qt::Variant#value
causes segfault, but it still seems better than doing it in the extension
since in that case we would have to figure out how to transfer the
objects from C++ to Ruby, and probably still end up getting the segfault.